+++ lib/speed_up.dart
@@ -5,5 +5,6 @@ export 'src/enum_extensions.dart';
 export 'src/funs.dart';
 export 'src/int_extensions.dart';
 export 'src/num_extensions.dart';
+export 'src/object_extensions.dart';
 export 'src/range_value.dart';
 export 'src/string_extensions.dart';