+++ lib/src/object_extensions.dart
@@ -1,0 +1,7 @@
+extension ObjectExtensions on Object? {
+  bool get isNull {
+    return this == null;
+  }
+
+  bool get isNotNull => !isNull;
+}