+++ test/collection_order_by_test.dart
@@ -5,19 +5,29 @@ import 'package:test/test.dart';
 import 'models.dart';
 
 void main() {
-  test('order by ASC', () {
-    const orders = [CoinOrder(2, 2), CoinOrder(1, 1)];
+  const orders = [
+    CoinOrder(2, 2),
+    CoinOrder(1, 1),
+    CoinOrder(3, 3),
+  ];
 
+  test('order by ASC', () {
     final ordered = orders.orderBy((x) => x.amount);
 
-    ordered.should.be(const [CoinOrder(1, 1), CoinOrder(2, 2)]);
+    ordered.should.be(const [
+      CoinOrder(1, 1),
+      CoinOrder(2, 2),
+      CoinOrder(3, 3),
+    ]);
   });
 
   test('order by DESC', () {
-    const orders = [CoinOrder(1, 1), CoinOrder(2, 2)];
-
     final ordered = orders.orderBy((x) => x.amount, desc: true);
 
-    ordered.should.be(const [CoinOrder(2, 2), CoinOrder(1, 1)]);
+    ordered.should.be(const [
+      CoinOrder(3, 3),
+      CoinOrder(2, 2),
+      CoinOrder(1, 1),
+    ]);
   });
 }